[#IOPID-1818] Refactor split configuration file
